Fixed Chicken and egg issue.
When you run Invoke-Command with the FilePath paramater, it executes the server inwhich you put in. On the server it will have the compiled executable which is great and all, but what if you need to invoke-command or ps-session into another server from there?
I welcome you to the chicken and the egg, I added a bit where it will compile the powershell file (the one that is used on the filepath parameter), it would base64 that file, and then add the base64 code to itself by appending it to the same file. Now, when executed with Invoke-Command, it will write two files to the second (or third, and so on) servers: The compiled executable, and the powershell script that makes it.